Move types for Zeitwerk
