--- app/types/interval_type.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-
-# This type maps between ActiveSupport::Duration and interval in PostgreSQL
-class IntervalType < ActiveRecord::Type::Value
-  def cast(value)
-    case value
-    when ::ActiveSupport::Duration
-      value
-    when ::String
-      ::ActiveSupport::Duration.parse(value)
-    else
-      super
-    end
-  rescue ActiveSupport::Duration::ISO8601Parser::ParsingError
-    nil
-  end
-
-  def serialize(value)
-    case value
-    when ::ActiveSupport::Duration
-      value.iso8601
-    else
-      super
-    end
-  end
-end
