+++ config/initializers/types.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-ActiveRecord::Type.register :interval, IntervalType, adapter: :postgresql
+require 'spina/conferences/types/interval_type'
+
+ActiveRecord::Type.register :interval, Spina::Conferences::IntervalType, adapter: :postgresql
 
 ActiveRecord::ConnectionAdapters::PostgreSQLAdapter.class_eval do
   alias_method :configure_connection_without_interval, :configure_connection
